% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_info.R
\name{stat_info}
\alias{stat_info}
\title{Function stat_info}
\usage{
stat_info(data)
}
\arguments{
\item{data}{is the input data to be summarised.}
}
\value{
the mean, median, range and variance of input data.
}
\description{
This function output some statistics information about the input data.
}
\examples{
stat_info(c(1,2,3))
}
